fix(long-to-int) :sparkles: fixed issue toInt() in long value upper t…

+++ app/src/main/java/com/rodrigo/deeplarva/routes/activity/CameraActivity.kt
@@ -65,7 +65,7 @@ class CameraActivity: AppCompatActivity() {
                 cameraProHW.updateISO(value)
             }
             override fun onUpdateShootSpeed(value: Int) {
-                cameraStore.updateShootSpeed(value)
+                cameraStore.updateShootSpeed(value.toLong())
                 cameraProHW.updateSpeed(value.toLong())
             }
         })
